Move ProjectDescription -> Paragraph
